requirements.txt
attrs
black
pytest
pytest-black
pytest-cov